requirements.txt
flask==2.0.1
pyngrok==5.0.5
python-dotenv==0.17.1
requests==2.25.1
tweepy==3.10.0
oauth2==1.9.0.post1
